Added comments to all methods and classes, improvement of null reference checking in methods

+++ src/net/azurewebsites/krystiankatafoniapp/controller/category/DeleteCategoryController.java
@@ -1,15 +1,21 @@
 package net.azurewebsites.krystiankatafoniapp.controller.category;
 
 import java.io.IOException;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import net.azurewebsites.krystiankatafoniapp.model.User;
 import net.azurewebsites.krystiankatafoniapp.service.CategoryService;
 
+/**
+ * Delete category controller 
+ * @author Krystian Katafoni
+ * @version 1.0
+ * @since 2017-06-05
+ */
 
 @WebServlet("/deleteCategory")
 public class DeleteCategoryController extends HttpServlet {
@@ -20,10 +26,15 @@ public class DeleteCategoryController extends HttpServlet {
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		
+		boolean result = false;
 		Long categoryId = Long.parseLong(request.getParameter("categoryId"));
 		CategoryService categoryService = new CategoryService();
-		boolean result=categoryService.deleteCategory(categoryId);
+		/*
+		 * Delete category
+		 */
+		if(categoryId!=null){
+			result=categoryService.deleteCategory(categoryId);
+		}
 		if(result==false){
 			request.getRequestDispatcher("WEB-INF/cannotDeleteCategory.jsp").forward(request,response);
 		}else{
