+++ src/net/azurewebsites/krystiankatafoniapp/controller/purchase/DeletePurchaseController.java
@@ -1,14 +1,20 @@
 package net.azurewebsites.krystiankatafoniapp.controller.purchase;
 
 import java.io.IOException;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import net.azurewebsites.krystiankatafoniapp.service.CategoryService;
 import net.azurewebsites.krystiankatafoniapp.service.PurchaseService;
+/**
+ * Delete purchase controller 
+ * @author Krystian Katafoni
+ * @version 1.0
+ * @since 2017-06-05
+ */
 
 @WebServlet("/deletePurchase")
 public class DeletePurchaseController extends HttpServlet {
@@ -18,10 +24,15 @@ public class DeletePurchaseController extends HttpServlet {
 	}
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
+		
 		Long purchaseId = Long.parseLong(request.getParameter("purchaseId"));
 		PurchaseService purchaseService = new PurchaseService();
-		purchaseService.deletePurchase(purchaseId);
+		/*
+		 * Delete purchase
+		 */
+		if(purchaseId!=null){
+			purchaseService.deletePurchase(purchaseId);
+		}
 		response.sendRedirect(request.getContextPath()+"/purchase");
 	}
